# Read optimizations. MemoryIO -> IO::Memory. Slice(UInt8) -> Bytes