--- a/shard.yml
+++ b/shard.yml
@@ -12,6 +12,6 @@ dependencies:
 development_dependencies:
   minitest:
     git: https://github.com/ysbaddaden/minitest.cr.git
-    version: ~> 0.3.1
+    version: ~> 0.3.5
 
 license: MIT