--- a/src/stumpy_png/filters.cr
+++ b/src/stumpy_png/filters.cr
@@ -1,28 +1,64 @@
 module StumpyPNG
   module Filter
-    def self.apply(scanline : Slice(UInt8), prior_scanline : Slice(UInt8), bpp, filter)
-      # Filter = NONE
-      return scanline if filter == 0
-
-      decoded = Slice(UInt8).new(scanline.size)
-
-      scanline.each_with_index do |byte, index|
-        prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
-        above = prior_scanline.empty? ? 0 : prior_scanline[index]
-        upper_left = (prior_scanline.empty? || (index - bpp) < 0) ? 0 : prior_scanline[index - bpp]
-
-        case filter
-        when 1 # Sub
+    def self.apply(scanline : Bytes, prior_scanline : Bytes, decoded : Bytes, bpp, filter) : Bytes
+      case filter
+      when 0 # None
+        return scanline
+      when 1 # Sub
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
           decoded[index] = byte + prior
-        when 2 # Up
-          decoded[index] = byte + above
-        when 3 # Average
-          decoded[index] = byte + ((prior.to_f + above.to_f) / 2).floor.to_u8
-        when 4 # Paeth
-          decoded[index] = byte + Utils.paeth_predictor(prior, above, upper_left)
-        else
-          raise "Unknown filter type #{filter}"
         end
+      when 2 # Up
+        scanline.each_with_index do |byte, index|
+          above = prior_scanline[index]
+          decoded[index] = byte + above
+        end
+      when 3 # Average
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
+          above = prior_scanline[index]
+          decoded[index] = byte + ((prior.to_f + above.to_f) / 2).floor.to_u8
+        end
+      when 4 # Paeth
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
+          above = prior_scanline[index]
+          upper_left = (index - bpp) < 0 ? 0 : prior_scanline[index - bpp]
+          decoded[index] = byte + Utils.paeth_predictor(prior, above, upper_left)
+        end
+      else
+        raise "Unknown filter type #{filter}"
+      end
+
+      decoded
+    end
+
+    def self.apply(scanline : Bytes, prior_scanline : Nil, decoded : Bytes, bpp, filter) : Bytes
+      case filter
+      when 0 # None
+        return scanline
+      when 1 # Sub
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
+          decoded[index] = byte + prior
+        end
+      when 2 # Up
+        scanline.each_with_index do |byte, index|
+          decoded[index] = byte
+        end
+      when 3 # Average
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
+          decoded[index] = byte + (prior.to_f / 2).floor.to_u8
+        end
+      when 4 # Paeth
+        scanline.each_with_index do |byte, index|
+          prior = (index - bpp) < 0 ? 0 : decoded[index - bpp]
+          decoded[index] = byte + Utils.paeth_predictor(prior, 0, 0)
+        end
+      else
+        raise "Unknown filter type #{filter}"
       end
 
       decoded