The biggest optimization is not using `each_slice` in `ColorTypes.decode`. `each_slice` returns an array for each n consecutive elements, so that was allocating way too many arrays.

Then I also moved the `case` expression of Filter.apply outside the loop, so that check has to be done one per scanline. I initially thought this isn't right, but other implementations (like oily_png) does that too. I guess there are many more places where this "moving outside the loop" can be done.

Then I also tried to reduce allocations in `to_canvas_none` by allocating just two slices and swapping them on every iteration, instead of allocating one per iteration.

Finally, I updated the code to Crystal 0.20.0 and did some other renames so code is shorter.

The result is not as good as I'd like, but since I don't know the underlying algorithms there isn't much I can do for now. I'm sure there's still a lot of room for optimizations. For example, while `NBitEnumerable` looks nice, I'm sure that not using that and computing pixel values in a loop must be faster (because now in color types we can only iterate one by one, accumulate into a slice, etc.). Probably having specialized code for each bit depth would be optimal, but this is just a guess.

Based on #7...

Before:

```
Read: 347.97ms (59.37%)
Transform: 16.57ms (2.83%)
Write: 221.58ms (37.8%)
Total: 586.12ms
```

After:

```
Read: 179.76ms (40.04%)
Transform: 17.27ms (3.85%)
Write: 251.97ms (56.12%)
Total: 449.0ms
```