@l3kn 2016-11-28
That's a huge improvement,
thank you very much!

I'll figure out why the automated tests are failing
and then merge this