@asterite 2016-11-28
Just in case what I said is a bit confusing, in oily_png there is [one function for each filter type](https://github.com/wvanbergen/oily_png/blob/master/ext/oily_png/png_decoding.c#L10-L54). There is also [one decoding function per combination of color mode and bit_depth](https://github.com/wvanbergen/oily_png/blob/master/ext/oily_png/png_decoding.c#L96-L254), so for example:

```c
oily_png_decode_scanline_grayscale_1bit
oily_png_decode_scanline_grayscale_2bit
oily_png_decode_scanline_grayscale_4bit
oily_png_decode_scanline_grayscale_8bit
oily_png_decode_scanline_grayscale_16bit
oily_png_decode_scanline_grayscale_alpha_8bit
oily_png_decode_scanline_grayscale_alpha_16bit
oily_png_decode_scanline_indexed_1bit
// and so on...
```

I believe this is optimal because doing a switch over the color type and bit depth in every loop iteration inside many other loop iterations will kill performance and disallow any inlining. So probably porting some of this code to Crystal will give a big performance boost (though I'm not sure, we'd have to try).