@asterite 2016-11-28
Sure, the code is this:

```cr
require "./src/stumpy_png"

def percent(label, time, total)
  puts "#{label}: #{time.total_milliseconds.round(2)}ms (#{(time.to_f * 100.0 / total.to_f).round(2)}%)"
end

time = Time.now
canvas = StumpyPNG.read("image.png")
read_time = Time.now - time

time = Time.now
(0...canvas.width).each do |x|
  (0...canvas.height).each do |y|
    color = canvas[x, y]
    g = 0_u32
    g = (g + color.r + color.g + color.b) / 3
    canvas[x, y] = StumpyPNG::RGBA.from_gray_n(g, 16)
  end
end
transform_time = Time.now - time

time = Time.now
StumpyPNG.write(canvas, "output.png")
write_time = Time.now - time

total_time = read_time + transform_time + write_time
percent "Read", read_time, total_time
percent "Transform", transform_time, total_time
percent "Write", write_time, total_time
puts "Total: #{total_time.total_milliseconds.round(2)}ms"
```