@asterite 2016-11-28
Oh, the tests were failing because now Crystal 0.20.0 is out and the old minitest didn't work with that, so I also just updated it to the latest version :-)