@l3kn 2016-11-28
I'm currently trying to rewrite the color decoding part
and you were right, removing the `NBitEnumerable`
improved the performance from ~220ms per read to ~205ms.

The code used to convert n-bit color values
to the 16bit values used by the `RGBA` class
was very slow, too (I tried make it work with bit-depths that are not a power of two).

After fixing all of the above problems,
the reading time is now at ~150ms.

I'll try to get these changes included as fast as possible.